fix: 修复codecheck问题

+++ openjiuwen/agent_builder/tune/base.py
@@ -19,7 +19,7 @@ class TuneConstant:
     DEFAULT_MAX_SAMPLED_EXAMPLE_NUM: int = 10
     DEFAULT_PARALLEL_NUM: int = 1
     DEFAULT_MAX_NUM_SAMPLE_ERROR_CASES: int = 10
-    DEFAULT_EARLY_STOP_SCORE: int = 1.0
+    DEFAULT_EARLY_STOP_SCORE: float = 1.0
 
     """optimizer parameters threshold constant"""
     MIN_ITERATION_NUM: int = 1
